y_newer_previous_attempt/pyproject.toml
[project]
name = "alex-szarka"
version = "0.1.0"
description = "Student Certification Readiness Multi-Agent System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework",
    "azure-ai-agentserver-agentframework",
    "azure-ai-projects",
    "azure-ai-agents",
    "azure-identity",
    "httpx",
    "openai",
    "python-dotenv",
    "starlette<1.0",
]

[tool.uv]
prerelease = "allow"
